refactored and uniformed SqSwingFileChooser to ScalaQuestComponents standards
